+++ packages/thumbprint-react/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 -   [Minor] Add `checkboxVerticalAlign` prop to `Checkbox` and `radioVerticalAlign` prop to `Radio` (#266)
 
+### Changed
+
+-   [Patch] Redesign the avatar with initials design to use dynamic colours. (#25)
+
 ## 2.1.0 - 2019-06-04
 
 ### Added