+++ packages/thumbprint-react/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 -   [Minor] Export a new component `ModalDefaultContentFullBleed` that allows full-bleed content for the `ModalDefault`. (#194)
 
+### Changed
+
+-   [Patch] Redesign the avatar with initials design to use dynamic colours. (#25)
+
 ## 2.1.0 - 2019-06-04
 
 ### Added