+++ packages/thumbprint-react/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+-   [Patch] Add small delay before showing tooltip after hovering to prevent flickering. (#158)
 -   [Patch] Redesign the avatar with initials design to use dynamic colours. (#25)
 
 ## 2.1.0 - 2019-06-04